+++ frontend/src/main/java/org/abs_models/backend/java/lib/runtime/AbstractAsyncCallRT.java
@@ -6,7 +6,6 @@ package org.abs_models.backend.java.lib.runtime;
 
 import org.abs_models.backend.java.JavaBackendException;
 import org.abs_models.backend.java.lib.types.ABSDataType;
-import org.abs_models.backend.java.lib.types.ABSRational;
 import org.abs_models.backend.java.lib.types.ABSRef;
 import org.apfloat.Aprational;
 
@@ -58,8 +57,7 @@ public abstract class AbstractAsyncCallRT<T extends ABSRef> extends AbstractAsyn
 
     private Aprational convertFromDuration(ABSDataType duration) {
         if (duration.getConstructorName().equals("Duration")) {
-            ABSRational rat = (ABSRational)duration.getArg(0);
-            return rat.toAprational();
+            return (Aprational)duration.getArg(0);
         } else if (duration.getConstructorName().equals("InfDuration")) {
             return new Aprational(-1);
         } else {